* Fix possible crash in fatal()

+++ src/main.c
@@ -192,7 +192,7 @@ void fatal(const char *s, int recoverable)
   if (my_port)
     listen_all(my_port, 1); /* close the listening port... */
 
-  if (conf.bot->localhub)
+  if (conf.bot && conf.bot->localhub)
     unlink(conf.localhub_socket);
 
   sdprintf(STR("Closing %d sockets"), dcc_total);
